Switch perlbug mail sender on VMS.
The VMS mail utility can't do attachments because it always adds
a blank line in front of any headers you add.  So use the Send
From File utility, which has been documented and supported for
any release in the last decade and was latently present before
that.

It takes the whole message verbatim just like sendmail, but also
needs the envelope prepended.