Storable: add SEEN0 macro
to shut up a compiler warning.

The SEEN(y,stash,i) macro expands to something like

    ...;
    if (stash)
       foo_NN(stash);

where foo_NN is a function that expects a non-null arg.

SEEN() is sometimes called as SEEN(sv,0,0) which of course expands to

    if (0)
       foo_NN(0);

which under gcc at least, emits the 'non-null' warning before the entire
block of code is optimised away.

So add a SEEN0() macro which handles the stash=0 case.