mark Perl_my_strftime with format attribute
mark this function with

    __attribute__format__null_ok__(__strftime__,pTHX_1,0)

so that compiler checks and warnings about strftime-style format args
can be checked.

Rather than adding new flag(s) to embed.fnc, I just enhanced the f flag
to treat it as strftime-style rather than printf if the function name
matches /strftime/. This was quicker, and we're unlikely to have many
such functions.