Remove tests of $File::Find::untaint_pattern.
This in effect reverts commit 5ea8618bf5cf20c62d2ccca6aca10d97e0945b89.
The effects of this package variable are not yet well enough understood to
write tests that pass consistently.  The block of tests removed was failing on
Win32 in all situations and on Linux in some.

For: RT #120711