WinCE makefile will now hypothetically build git_version.h
The dependency on $(CONFIGPM) is for git_version.h to exist early enough
to be copied to the CORE dir. In real life, git_version.h still is built
by a Desktop build since WinCE doesn't yet support being built with a
system Perl. This commit is a step in getting WinCE to build without
needing a Desktop build first and for better integration with Castaway's
Cross changes in the future. The changes in this commit were tested with
other unpublished changes that tried to do a WinCE build with a system
Perl.