lib/locale.t: Pass AIX failures
Some AIX locales call a no-break space "graphic".  Rather than fail AIX
for this, this commit adopts the same approach used for MSWin, which
similarly has most locales containing some errors.  And that is to pass
it if any locales at all are correct.  Ideally the vendors wouldn't make
broken locales, but this allows us to test for full legality on other
systems.