pp_sys.c: Remove null checks from pp_sockpair
There is actually no way for nulls to reach this code.

Nulls on the stack only happen with pp_coreargs, and only with
ops that have optional arguments, of which socketpair is not one.

GvIOn uses gv_add_by_type, which adds a new IO if there is not
already one, so it will never return null if the GV is not null.